--- Assets/Failsafe/Scripts/ItemSystem/IremTags/ItemRarity.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 1629bfd0c4df03044b4117e3f691e46e
