+++ Assets/Failsafe/Scripts/ItemSystem/Items.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3d6d3ea4f04a8054d8a83aa63efe757a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
