+++ Assets/Failsafe/Scripts/ItemSystem/Items/IUsable.cs
@@ -1,0 +1,4 @@
+public interface IUsable
+{
+    public void Use();
+}
