+++ Assets/Failsafe/Scripts/ItemSystem/Items/IUsable.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: c601918827a173e4faffeb7e79809346
