+++ Assets/Failsafe/Scripts/ItemSystem/Items/ItemEnums.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: d86add453438bcb4a9023585e52475c2
