+++ Assets/Failsafe/Scripts/ItemSystem/Items/UsableItem.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: dda2f5a5d2a917d4682f5848ae185366
